Moved all typed data in RAM related code into its own file
